refactor: TodoList 구조 변경
