+++ src/components/TodoListBody/TodoListBody.tsx
@@ -1,5 +1,5 @@
 import { styled } from 'styled-components'
-import { useTodoContext } from '../../TodoProvider'
+import { useTodoContext } from '../TodoList/TodoList'
 import TodoItem from '../TodoItem/TodoItem'
 
 function TodoListBody() {
